+++ minijinja/src/vm/state.rs
@@ -468,6 +468,10 @@ impl<'template, 'env> BlockStack<'template, 'env> {
         self.instructions.get(self.depth).copied().unwrap()
     }
 
+    pub fn len(&self) -> usize {
+        self.instructions.len()
+    }
+
     pub fn push(&mut self) -> bool {
         if self.depth + 1 < self.instructions.len() {
             self.depth += 1;